CSN: CSN-asimcore-337
Sync of Intel commit 3374

mgambhir    Date: Mon Jun  6 14:55:24 EDT 2011  CSN: CSN-simcore-3372

- fixed warnings when built with gcc 4.3.2
  - char * variable when passed a const string threw a warning with gcc
    4.3.2. To fix this a lot of function prototypes were changed to take
    a const char* parameter instead of a char*
  - some conversion warnings in libdraldb like uint32 to uint16 conversion
    were turned off by removing -Wconversion flag
  - some data type conversion warnings were removed by using
    static_cast<T> and const_cast<T>

+++ lib/libdraldb/include/asim/ItemHandler.h
@@ -397,7 +397,7 @@ ItemHandler::getFormatedTagValue()
     UINT64 value;
     INT32 tagId = itemtagheap->getTagId(this);
     value = itemtagheap->getValue(this);
-    return tagdescvec->getFormatedTagValue(tagId, value);
+    return tagdescvec->getFormatedTagValue(static_cast<UINT16>(tagId), value);
 }
 
 /**
